[refactor] listItem Component 분리
